feat!: move options to parent namespace because the Middleware namespace is no longer relevant. Furthermore, ensure
